refactor(web): extract AppSidebar from App.tsx
Mechanical move of the ~250-line <aside> block into
components/app/AppSidebar.tsx, with AgentNavList / ProjectNavList
split out as internal subcomponents to keep the prop surface bounded.
getProjectGroupIdentity moves to lib/projects.ts as it is shared
between App.tsx and the sidebar.

Issue: CS-18 (step 1/3)
