+++ apps/web/src/App.tsx
@@ -3,8 +3,7 @@ declare const __APP_VERSION__: string;
 import { useCallback, useEffect, useEffectEvent, useMemo, useState, type ReactNode } from "react";
 import { PanelLeftClose, PanelLeftOpen } from "lucide-react";
 import { Link, useLocation, useNavigate } from "react-router-dom";
-import { ModelConfig } from "./config";
-import type { SessionHead, ProjectGroup } from "./lib/api";
+import type { SessionHead } from "./lib/api";
 import { logClientEvent } from "./lib/api";
 import { SessionDetail } from "./components/SessionDetail";
 import { SessionDetailSkeleton } from "./components/SessionDetailSkeleton";
@@ -12,12 +11,9 @@ import { DetailLanding, type LandingAgentItem } from "./components/DetailLanding
 import { Dashboard } from "./components/Dashboard";
 import { ProjectDashboardView, ProjectsOverview } from "./components/Projects";
 import { ErrorBoundary } from "./components/ErrorBoundary";
-import { BookmarkButton } from "./components/BookmarkButton";
 import { CopyResumeButton } from "./components/CopyResumeButton";
-import { SessionTreeSidebar } from "./components/SessionTreeSidebar";
 import { RenderProfiler } from "./components/RenderProfiler";
 import { SmartTagChips } from "./components/SmartTagChips";
-import { getSessionBookmarkKey } from "./lib/bookmarks";
 import { parseViewState } from "./lib/view-state";
 import { useScanStatus } from "./hooks/useScanStatus";
 import { useSessionDetail } from "./hooks/useSessionDetail";
@@ -31,19 +27,17 @@ import { useSessions } from "./hooks/useSessions";
 import { useProjects } from "./hooks/useProjects";
 import { useInitialLoad } from "./hooks/useInitialLoad";
 import { useLiveSync } from "./hooks/useLiveSync";
-import { BrowseByToggle } from "./components/app/BrowseByToggle";
-import { SidebarFlatSessionList } from "./components/app/SidebarFlatSessionList";
+import { AppSidebar } from "./components/app/AppSidebar";
 import { SearchResultsPanel } from "./components/app/SearchResultsPanel";
 import { type BrowseBy, type SearchProjectOption } from "./components/app/types";
-import {
-  formatAgentScanProgress,
-  formatScanStatusLabel,
-  formatSearchSubtitle,
-  formatWindowLabel,
-  getAgentDisplayCount,
-} from "./lib/scan-format";
+import { formatScanStatusLabel, formatSearchSubtitle, formatWindowLabel } from "./lib/scan-format";
 import { formatRelativeTime } from "./lib/format";
-import { getProjectIdentityKey, getProjectPath, type ProjectRouteIdentity } from "./lib/projects";
+import {
+  getProjectGroupIdentity,
+  getProjectIdentityKey,
+  getProjectPath,
+  type ProjectRouteIdentity,
+} from "./lib/projects";
 import {
   buildSessionIndexes,
   buildSidebarSessionLookup,
@@ -51,10 +45,6 @@ import {
   getSessionAgentKey,
   getSessionRouteKey,
 } from "./lib/session-indexes";
-
-function getProjectGroupIdentity(project: ProjectGroup): ProjectRouteIdentity {
-  return { kind: project.identityKind, key: project.identityKey };
-}
 
 interface BreadcrumbItem {
   label: string;
@@ -947,263 +937,28 @@ export default function App() {
       </header>
 
       <div className="flex min-h-0 flex-1">
-        <aside
-          className={`w-64 shrink-0 flex-col border-r border-[var(--console-border)] bg-[var(--console-sidebar-bg)] ${
-            sidebarCollapsed ? "hidden" : "hidden lg:flex"
-          }`}
-        >
-          <div className="console-scrollbar flex-1 space-y-8 overflow-y-auto px-4 py-6">
-            <section>
-              <h3 className="console-mono mb-3 text-xs font-bold uppercase text-[var(--console-text)]">
-                BROWSE BY
-              </h3>
-              <BrowseByToggle
-                value={browseBy}
-                onChange={changeBrowseBy}
-                projectsDisabled={isScanActive}
-              />
-            </section>
-
-            <section>
-              <h3 className="console-mono mb-3 text-xs font-bold uppercase text-[var(--console-text)]">
-                NAVIGATION
-              </h3>
-              <ul
-                className={`space-y-1 ${
-                  browseBy === "projects"
-                    ? "console-scrollbar max-h-[min(280px,calc(100vh-440px))] overflow-y-auto pr-1"
-                    : ""
-                }`}
-              >
-                <li>
-                  <Link
-                    to={browseBy === "projects" ? "/projects" : "/"}
-                    className={`flex items-center gap-2 rounded-sm border px-3 py-1.5 text-left transition-colors ${
-                      (browseBy === "agents" && viewState.mode === "root") ||
-                      (browseBy === "projects" && viewState.mode === "projects")
-                        ? "border-[var(--console-border-strong)] bg-white text-[var(--console-text)]"
-                        : "border-transparent text-[var(--console-muted)] hover:border-[var(--console-border)] hover:bg-[var(--console-surface-muted)]"
-                    }`}
-                  >
-                    <img src="/logo.svg?v=3" alt="Dashboard" className="size-3.5 rounded-[2px]" />
-                    <span className="console-mono line-clamp-1 flex-1 text-xs">Dashboard</span>
-                  </Link>
-                </li>
-                {browseBy === "agents"
-                  ? agents.map((agent) => {
-                      const key = agent.name.toLowerCase();
-                      const isSelected = key === activeAgentKey;
-                      const config = ModelConfig.agents[key];
-                      const agentProgress = formatAgentScanProgress(scanStatus, agent.name);
-                      const disabled = isScanActive && agentProgress !== null;
-                      const className = `ml-4 flex items-center gap-2 rounded-sm border px-3 py-1.5 text-left transition-colors ${
-                        disabled
-                          ? "cursor-not-allowed border-transparent text-[var(--console-muted)] opacity-50"
-                          : isSelected
-                            ? "border-[var(--console-border-strong)] bg-white text-[var(--console-text)]"
-                            : "border-transparent text-[var(--console-muted)] hover:border-[var(--console-border)] hover:bg-[var(--console-surface-muted)]"
-                      }`;
-                      const content = (
-                        <>
-                          {config?.icon && (
-                            <img
-                              src={config.icon}
-                              alt={agent.displayName}
-                              className="size-3.5 object-contain"
-                            />
-                          )}
-                          <span className="console-mono line-clamp-1 flex-1 text-xs">
-                            {agent.displayName}
-                          </span>
-                          <span className="console-mono text-[11px] text-[var(--console-muted)]">
-                            {agentProgress ??
-                              getAgentDisplayCount(scanStatus, agent.name, agent.count)}
-                          </span>
-                        </>
-                      );
-                      return (
-                        <li key={agent.name}>
-                          {disabled ? (
-                            <span
-                              className={className}
-                              title="Available after this agent scan completes"
-                            >
-                              {content}
-                            </span>
-                          ) : (
-                            <Link to={`/${key}`} className={className}>
-                              {content}
-                            </Link>
-                          )}
-                          {agentProgress ? (
-                            <span className="ml-4 mt-1 block h-1 overflow-hidden rounded-sm bg-[var(--console-surface-muted)]">
-                              <span
-                                className="block h-full bg-[var(--console-accent)]"
-                                style={{
-                                  width: `${
-                                    scanStatus?.agentStatuses[agent.name]?.total
-                                      ? Math.round(
-                                          ((scanStatus.agentStatuses[agent.name]?.processed ?? 0) /
-                                            scanStatus.agentStatuses[agent.name]!.total!) *
-                                            100,
-                                        )
-                                      : 8
-                                  }%`,
-                                }}
-                              />
-                            </span>
-                          ) : null}
-                        </li>
-                      );
-                    })
-                  : projects.map((project) => {
-                      const projectIdentity = getProjectGroupIdentity(project);
-                      const isSelected =
-                        selectedProjectNavigationId === getProjectIdentityKey(projectIdentity);
-                      return (
-                        <li key={`${project.identityKind}:${project.identityKey}`}>
-                          <Link
-                            to={getProjectPath(projectIdentity)}
-                            onClick={() => setSelectedProjectIdentity(projectIdentity)}
-                            className={`ml-4 flex min-w-0 items-center gap-2 rounded-sm border px-3 py-1.5 text-left transition-colors ${
-                              isSelected
-                                ? "border-[var(--console-border-strong)] bg-white text-[var(--console-text)]"
-                                : "border-transparent text-[var(--console-muted)] hover:border-[var(--console-border)] hover:bg-[var(--console-surface-muted)]"
-                            }`}
-                          >
-                            <span className="console-mono min-w-0 flex-1 truncate text-xs">
-                              {project.displayName}
-                            </span>
-                            <span className="console-mono shrink-0 text-[11px] text-[var(--console-muted)]">
-                              {project.sessionCount}
-                            </span>
-                          </Link>
-                        </li>
-                      );
-                    })}
-                {browseBy === "agents" && agents.length === 0 && !loading ? (
-                  <li>
-                    <span className="console-mono block rounded-sm px-3 py-1.5 text-xs text-[var(--console-muted)]">
-                      {scanStatus?.active ? "Scanning agents..." : "No agents found"}
-                    </span>
-                  </li>
-                ) : null}
-                {browseBy === "projects" && projects.length === 0 && !loading ? (
-                  <li>
-                    <span className="console-mono block rounded-sm px-3 py-1.5 text-xs text-[var(--console-muted)]">
-                      {scanStatus?.active ? "Scanning projects..." : "No projects found"}
-                    </span>
-                  </li>
-                ) : null}
-              </ul>
-            </section>
-
-            <section>
-              <h3 className="console-mono mb-3 text-xs font-bold uppercase text-[var(--console-text)]">
-                BOOKMARKS
-              </h3>
-              {bookmarkedSessions.length === 0 ? (
-                <span className="console-mono block rounded-sm px-3 py-1.5 text-xs text-[var(--console-muted)]">
-                  No bookmarks yet
-                </span>
-              ) : (
-                <ul className="space-y-1">
-                  {bookmarkedSessions.map((session) => {
-                    const isActive =
-                      viewState.mode === "session" &&
-                      viewState.activeAgentKey === session.agentKey &&
-                      viewState.activeSessionSlug === session.sessionId;
-                    const agent = ModelConfig.agents[session.agentKey];
-                    return (
-                      <li key={getSessionBookmarkKey(session.agentKey, session.sessionId)}>
-                        <div
-                          className={`flex items-start gap-2 rounded-sm border px-2 py-1.5 transition-colors ${
-                            isActive
-                              ? "border-[var(--console-border-strong)] bg-white text-[var(--console-text)]"
-                              : "border-transparent text-[var(--console-muted)] hover:border-[var(--console-border)] hover:bg-[var(--console-surface-muted)]"
-                          }`}
-                        >
-                          <Link
-                            to={`/${session.fullPath}`}
-                            className="flex min-w-0 flex-1 items-start gap-2"
-                          >
-                            {agent?.icon ? (
-                              <img
-                                src={agent.icon}
-                                alt={agent.name}
-                                className="mt-0.5 size-3.5 shrink-0 object-contain"
-                              />
-                            ) : null}
-                            <div className="min-w-0 flex-1">
-                              <span className="console-mono line-clamp-1 block text-xs">
-                                {session.title}
-                              </span>
-                              <span className="console-mono mt-0.5 line-clamp-1 block text-[10px] text-[var(--console-muted)]">
-                                {agent?.name ?? session.agentKey}
-                              </span>
-                            </div>
-                          </Link>
-                          <BookmarkButton active onToggle={() => toggleBookmark(session)} />
-                        </div>
-                      </li>
-                    );
-                  })}
-                </ul>
-              )}
-            </section>
-
-            <section>
-              <h3 className="console-mono mb-3 text-xs font-bold uppercase text-[var(--console-text)]">
-                SESSIONS
-                {sidebarSessions.length > 0 ? (
-                  <span className="ml-2 text-[10px] font-normal text-[var(--console-muted)]">
-                    Navigate j k · Open Enter
-                  </span>
-                ) : null}
-              </h3>
-              {browseBy === "agents" && !activeAgentKey ? (
-                <span className="console-mono block rounded-sm px-3 py-1.5 text-xs text-[var(--console-muted)]">
-                  Select an agent
-                </span>
-              ) : browseBy === "projects" && !selectedProjectNavigationId ? (
-                <span className="console-mono block rounded-sm px-3 py-1.5 text-xs text-[var(--console-muted)]">
-                  Select a project
-                </span>
-              ) : sidebarSessions.length === 0 ? (
-                <span className="console-mono block rounded-sm px-3 py-1.5 text-xs text-[var(--console-muted)]">
-                  {scanStatus?.active ? "Scanning sessions..." : "No sessions yet"}
-                </span>
-              ) : browseBy === "projects" ? (
-                <SidebarFlatSessionList
-                  sessions={sidebarSessions}
-                  activeSessionId={
-                    viewState.mode === "session" ? viewState.activeSessionSlug : null
-                  }
-                  selectedSessionId={selectedSidebarSessionId}
-                  bookmarkedSessionIds={bookmarkedSidebarSessionIds}
-                  onSelectSession={handleSelectFlatSidebarSession}
-                  onToggleBookmark={handleToggleSidebarSessionBookmark}
-                />
-              ) : (
-                <RenderProfiler
-                  id="SessionTreeSidebar"
-                  detail={{ sessions: sidebarSessions.length }}
-                >
-                  <SessionTreeSidebar
-                    sessions={sidebarSessions}
-                    activeSessionId={
-                      viewState.mode === "session" ? viewState.activeSessionSlug : null
-                    }
-                    selectedSessionId={selectedSidebarSessionId}
-                    onSelectSession={handleSelectTreeSidebarSession}
-                    bookmarkedSessionIds={bookmarkedSidebarSessionIds}
-                    onToggleBookmark={handleToggleSidebarSessionBookmark}
-                  />
-                </RenderProfiler>
-              )}
-            </section>
-          </div>
-        </aside>
+        <AppSidebar
+          sidebarCollapsed={sidebarCollapsed}
+          browseBy={browseBy}
+          onChangeBrowseBy={changeBrowseBy}
+          isScanActive={isScanActive}
+          viewState={viewState}
+          agents={agents}
+          activeAgentKey={activeAgentKey}
+          scanStatus={scanStatus}
+          projects={projects}
+          selectedProjectNavigationId={selectedProjectNavigationId}
+          onSelectProject={setSelectedProjectIdentity}
+          loading={loading}
+          bookmarkedSessions={bookmarkedSessions}
+          onToggleBookmark={toggleBookmark}
+          sidebarSessions={sidebarSessions}
+          selectedSidebarSessionId={selectedSidebarSessionId}
+          bookmarkedSidebarSessionIds={bookmarkedSidebarSessionIds}
+          onSelectFlatSidebarSession={handleSelectFlatSidebarSession}
+          onToggleSidebarSessionBookmark={handleToggleSidebarSessionBookmark}
+          onSelectTreeSidebarSession={handleSelectTreeSidebarSession}
+        />
 
         <main className="flex min-w-0 flex-1 flex-col">
           <section className="shrink-0 border-b border-[var(--console-border)] bg-white/70 px-4 py-4 backdrop-blur-sm md:px-8">
